requirements.txt
librosa~=0.8.1

matplotlib~=3.4.1
scikit-learn~=0.24.2
numpy~=1.20.2
mysql-connector-python~=8.0.25
pandas~=1.2.4